Use the format for source locations. (find-function-locations): New function.  Replaces function-source-location-for-emacs.  Returns a list of source-locations.
(resolve-note-location): Renamed from resolve-location.  Simplified.
(brief-compiler-message-for-emacs): Print the source context (that's
the thing after ==>).

(who-xxxx): Take strings, not symbols, as arguments.
(function-callees, function-callers): Use the same format as the
who-xxx functions.  Support for byte-compiled stuff.

(code-location-stream-position): Try to be clever is the source path
doesn't match the form.

(call-with-debugging-environment): Bind *print-readably* to nil.