Major refactoring.
Restructured and documented starting at the top of the file and going
down through the networking protocol.

Some highlights:

(slime-rex-continuations): Now connection-local.  Mostly resectioning
and reordering definitions to try and improve readability.

(slime-get-temp-buffer-create): New utility function to popup a
temporary buffer that automatically has a binding on `q' to
intelligently restore window configuration. Handy, but currently not
applicable to all of our temporary buffers.

(slime-with-chosen-connection): Removed this macro. Consequently the
compilation commands no longer prompt for which connection to use when
given a prefix argument. `slime-switch-to-output-buffer' still works
like that, but for other cases I think the connection-list buffer is
sufficient.

(slime-eval-async): New arglist: (form &optional cont pkg). If the
continuation is unspecified then the evaluation result is ignored, and
if the package is unspecified then (slime-buffer-package) is used.

(slime-eval): Package arg now defaults to (slime-buffer-package).

(slime-current-package): New name for (slime-buffer-package). No more
caching: returns the buffer-local `slime-buffer-package' if set,
otherwise searches for an `in-package' form.  A consequence of
non-caching is that the package name doesn't appear in the modeline
anymore. The simplification is worthwhile in my opinion.