(slime-dispatch-event): Accept stepping flag.
(slime-space): Call slime-message in the right buffer, so that
after-command hooks are added in the right buffer.  Reported by Juho
Snellman.

(sldb-setup): Don't query when entering a recursive edit.

(sldb-exit): Don't kill the buffer if we are in stepping mode.

(slime-inspector-insert-ispec): New function.
(slime-open-inspector): Use it.
(slime-inspector-operate-on-point): Simplified.

(test interactive-eval): Fix test case.