(slime-indent-and-complete-symbol): Echo the arglist if there's no symbol before point.  Don't complete after parens. (slime-echo-arglist): Factorized from slime-space. (slime-space): Use it.
(slime-repl-history-replace): Add argument to clear the input at the
end of the history.

(slime-net-coding-system): Emacs does funky encoding for
`raw-text-unix' use `binary' instead.
(slime-safe-encoding-p): New function.
(slime-net-send): Use it and don't try to send stuff which can't be
decoded by Lisp.

(slime-inferior-lisp-program-history): XEmacs compatibility: declare
it as a variable.

(slime-xref-mode): In Emacs 21, set delayed-mode-hooks to nil because
we don't want to run the lisp-mode-hook.