(slime, slime-start): Introduce a separate function for the non-interactive case.  `slime-start' takes lots of keyword arguments and `slime' is reserved for interactive use. (slime-read-interactive-args): New function.
(slime-maybe-start-lisp, slime-inferior-lisp)
(slime-start-swank-server): Pass all argumenets needed to start the
subprocess as a property list.  Also store this list in a buffer-local
var in the inferior-lisp buffer, so that we can cleanly restart the
process.

(slime-registered-lisp-implementations): Change the format and
document it.  M-- M-x slime can now be used select a registered
implemetantion.

(slime-symbolic-lisp-name): Deleted. And updated all the functions
which pass it along.

(slime-set-connection-info): Use the new format.

(slime-output-buffer): Don't re-initialize buffer-local variables if
the buffer already exists.  This saves the history. From Juho
Snellman.